WIP on solution package refactoring;
