+++ solvis/fault_system_solution/__init__.py
@@ -1,0 +1,2 @@
+from .fault_system_solution import FaultSystemSolution
+from .fault_system_solution_model import FaultSystemSolutionModel
