+++ solvis/fault_system_solution/fault_system_solution_file.py
@@ -5,7 +5,7 @@ from typing import TYPE_CHECKING, Optional
 import geopandas as gpd
 import pandas as pd
 
-from .inversion_solution_file import InversionSolutionFile, data_to_zip_direct
+from solvis.inversion_solution.inversion_solution_file import InversionSolutionFile, data_to_zip_direct
 
 if TYPE_CHECKING:
     from pandera.typing import DataFrame
