+++ solvis/fault_system_solution/fault_system_solution_model.py
@@ -9,9 +9,10 @@ from typing import TYPE_CHECKING, Optional, cast
 
 from pandera.typing import DataFrame
 
-from .dataframe_models import RuptureSectionSchema
+from solvis.inversion_solution.dataframe_models import RuptureSectionSchema
+from solvis.inversion_solution.inversion_solution_model import InversionSolutionModel
+
 from .fault_system_solution_file import FaultSystemSolutionFile
-from .inversion_solution_model import InversionSolutionModel
 
 # from .typing import AggregateSolutionProtocol
