+++ solvis/inversion_solution/__init__.py
@@ -26,8 +26,6 @@ Example:
 """
 
 from .composite_solution import CompositeSolution
-from .fault_system_solution import FaultSystemSolution
-from .fault_system_solution_model import FaultSystemSolutionModel
 from .inversion_solution import InversionSolution
 
 """
